+++ providers/index.js
@@ -1,14 +1,19 @@
 const { languages } = require('vscode');
 
 function register(provider) {
-  return languages.registerCompletionItemProvider('javascript', provider, '.');
+  return languages.registerCompletionItemProvider(
+    { scheme: 'file', language: 'javascript' },
+    provider,
+    '.',
+  );
 }
 
 module.exports = [
   register(require('./modules')),
   register(require('./Users')),
+  register(require('./wixCRM')),
   register(require('./wixUsers')),
-  register(require('./wixUsersBackend')),
+  register(require('./wixMarketing')),
   register(require('./wixWindow')),
   register(require('./wixStorage')),
   register(require('./wixData')),