+++ providers/wixCRM.js
@@ -4,19 +4,22 @@ const { createCompletionList, isBackend } = require('./util');
 const K = vs.CompletionItemKind;
 
 const listFronend = createCompletionList([
-  { name: 'getProductOptionsAvailability', kind: K.Method },
+  { name: 'createContact', kind: K.Method },
+  { name: 'emailContact', kind: K.Method },
 ]);
 
 const listBackend = createCompletionList([
-  { name: 'getCurrentCart', kind: K.Method },
-  { name: 'getProductOptionsAvailability', kind: K.Method },
+  { name: 'createContact', kind: K.Method },
+  { name: 'emailContact', kind: K.Method },
+  { name: 'getContactById', kind: K.Method },
+  { name: 'updateContact', kind: K.Method },
 ]);
 
 module.exports = {
   provideCompletionItems(doc, position) {
     const prefix = doc.lineAt(position).text.substr(0, position.character);
 
-    if (prefix.endsWith('wixStores.')) {
+    if (prefix.endsWith('wixCRM.')) {
       return isBackend(doc.uri.path) ? listBackend : listFronend;
     }
   },